Rename launcher/verifier/grpc to launcher/verifier/grpcclient
