+++ launcher/agent/agent_test.go
@@ -9,13 +9,13 @@ import (
 
 	"github.com/google/go-tpm-tools/client"
 	"github.com/google/go-tpm-tools/internal/test"
-	grpcclient "github.com/google/go-tpm-tools/launcher/verifier/grpc"
-	"github.com/google/go-tpm-tools/launcher/verifier/grpc/service"
+	"github.com/google/go-tpm-tools/launcher/verifier/grpcclient"
+	"github.com/google/go-tpm-tools/launcher/verifier/grpcclient/service"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/grpc/test/bufconn"
 
-	servgrpc "github.com/google/go-tpm-tools/launcher/verifier/grpc/proto/attestation_verifier/v0"
+	servgrpc "github.com/google/go-tpm-tools/launcher/verifier/grpcclient/proto/attestation_verifier/v0"
 )
 
 func TestAttest(t *testing.T) {
@@ -41,8 +41,7 @@ func TestAttest(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed to connect to attestation service: %v", err)
 	}
-	pbClient := servgrpc.NewAttestationVerifierClient(conn)
-	verifierClient := grpcclient.NewClient(pbClient, log.Default())
+	verifierClient := grpcclient.NewClient(conn, log.Default())
 	// Cannot test a GCE key on the simulator.
 	agent := CreateAttestationAgent(tpm, client.AttestationKeyECC, verifierClient, placeholderFetcher)
