+++ launcher/container_runner.go
@@ -23,8 +23,7 @@ import (
 	"github.com/google/go-tpm-tools/client"
 	"github.com/google/go-tpm-tools/launcher/agent"
 	"github.com/google/go-tpm-tools/launcher/spec"
-	grpcclient "github.com/google/go-tpm-tools/launcher/verifier/grpc"
-	servpb "github.com/google/go-tpm-tools/launcher/verifier/grpc/proto/attestation_verifier/v0"
+	"github.com/google/go-tpm-tools/launcher/verifier/grpcclient"
 	v1 "github.com/opencontainers/image-spec/specs-go/v1"
 	specs "github.com/opencontainers/runtime-spec/specs-go"
 	"golang.org/x/oauth2"
@@ -168,8 +167,7 @@ func NewRunner(ctx context.Context, cdClient *containerd.Client, token oauth2.To
 	if err != nil {
 		return nil, fmt.Errorf("failed to open connection to attestation service: %v", err)
 	}
-	pbClient := servpb.NewAttestationVerifierClient(conn)
-	verifierClient := grpcclient.NewClient(pbClient, logger)
+	verifierClient := grpcclient.NewClient(conn, logger)
 
 	// Fetch ID token with specific audience.
 	// See https://cloud.google.com/functions/docs/securing/authenticating#functions-bearer-token-example-go.
